more CustomApplicationTypeSchemaUtils tests
